Changed to use short int instead of int for gameBoard

+++ MankalahAI/gameState.h
@@ -8,13 +8,13 @@
 
 // board constants
 const int holes = 7;
-const int seeds = 7;
+const short int seeds = 7;
 const int boardSize = 2 * holes + 2;
-const int totalSeeds = 2 * holes * seeds;
-const std::array<int, boardSize> defaultBoard =
+const short int totalSeeds = 2 * holes * seeds;
+const std::array<short int, boardSize> defaultBoard =
 	{ 7,7,7,7,7,7,7,0,7,7,7,7,7,7,7,0 };
 
-typedef std::array<int, boardSize> gameBoard;
+typedef std::array<short int, boardSize> gameBoard;
 std::ostream& operator<<(std::ostream& os, const gameBoard& board);
 
 class gameState {
@@ -31,7 +31,7 @@ public:
 	float getValue(int player) const;
 	std::set<int> movesAvailable() const;
 	std::set<int> movesAvailable(int player) const;
-	std::array<int, 2> scores() const;
+	std::array<short int, 2> scores() const;
 	gameState moveResult(int pos) const;
 	bool isGameOver() const;
 	int getCurrentPlayer() const;
